# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DevLog+",
    "description": "Single-user developer journal for technical learning and skill maintenance.\n\nDevLog+ has two engines:\n\n- **Learning engine** — builds a Knowledge Profile from journal entries, quizzes, and feedback. Understands what you know, where you're weak, and what to explore next.\n- **Practice engine** — generates weekly Go micro-projects calibrated to your practical level.\n\n## Core workflow\n\n1. **Journal** — write technical entries (typed or dictated) about what you learn.\n2. **Profile** — the system builds and maintains a Knowledge Profile nightly.\n3. **Quiz** — weekly free-text quizzes probe your understanding.\n4. **Readings** — curated reading recommendations from your allowlisted domains.\n5. **Projects** — weekly Go micro-projects with bugs, features, and refactors.\n6. **Feedback** — thumbs-up/down reactions and feedforward notes steer future content.\n7. **Triage** — items the system can't resolve are surfaced for your clarification.\n",
    "license": {
      "name": "Private — single-user local application"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/readings/recommendations": {
      "get": {
        "tags": [
          "readings"
        ],
        "summary": "List reading recommendations",
        "description": "Return reading recommendations ordered by batch date (most recent first).\n\nRecommendations are generated weekly from the Knowledge Profile and limited\nto domains on the user's allowlist.  Each batch typically contains 3–5 links.",
        "operationId": "list_recommendations_api_v1_readings_recommendations_get",
        "parameters": [
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Number of recommendations to skip",
              "default": 0,
              "title": "Offset"
            },
            "description": "Number of recommendations to skip"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Maximum recommendations to return",
              "default": 20,
              "title": "Limit"
            },
            "description": "Maximum recommendations to return"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of reading recommendations, most recent batch first",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReadingRecommendationResponse"
                  },
                  "title": "Response List Recommendations Api V1 Readings Recommendations Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/readings/allowlist": {
      "get": {
        "tags": [
          "readings"
        ],
        "summary": "List allowlisted domains",
        "description": "Return every domain on the reading allowlist.\n\nThe allowlist controls which domains the reading-generation pipeline may\nlink to.  Default entries are pre-populated on first run.",
        "operationId": "list_allowlist_api_v1_readings_allowlist_get",
        "responses": {
          "200": {
            "description": "All domains currently on the reading allowlist",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AllowlistEntryResponse"
                  },
                  "type": "array",
                  "title": "Response List Allowlist Api V1 Readings Allowlist Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "readings"
        ],
        "summary": "Add allowlist domain",
        "description": "Add a new domain to the reading allowlist.\n\nOnce added, the reading-generation pipeline may include links from this\ndomain in future weekly recommendations.",
        "operationId": "add_allowlist_entry_api_v1_readings_allowlist_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AllowlistEntryCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "The newly created allowlist entry",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AllowlistEntryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error — domain and name are required"
          }
        }
      }
    },
    "/api/v1/readings/allowlist/{entry_id}": {
      "put": {
        "tags": [
          "readings"
        ],
        "summary": "Update allowlist entry",
        "description": "Update the name or description of an existing allowlist entry.",
        "operationId": "update_allowlist_entry_api_v1_readings_allowlist__entry_id__put",
        "parameters": [
          {
            "name": "entry_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Entry Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AllowlistEntryUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated allowlist entry",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AllowlistEntryResponse"
                }
              }
            }
          },
          "404": {
            "description": "Allowlist entry not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "readings"
        ],
        "summary": "Remove allowlist domain",
        "description": "Remove a domain from the reading allowlist.\n\nFuture reading recommendations will no longer include links from this\ndomain.  Existing recommendations are not affected.",
        "operationId": "delete_allowlist_entry_api_v1_readings_allowlist__entry_id__delete",
        "parameters": [
          {
            "name": "entry_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Entry Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Confirmation message",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "404": {
            "description": "Allowlist entry not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/settings/{key}": {
      "get": {
        "tags": [
          "settings"
        ],
        "summary": "Get a setting by key",
        "description": "Retrieve a single setting by its unique key.",
        "operationId": "get_setting_api_v1_settings__key__get",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The requested setting",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingResponse"
                }
              }
            }
          },
          "404": {
            "description": "Setting not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "settings"
        ],
        "summary": "Create or update a setting",
        "description": "Create a new setting or update an existing one.\n\nThe `value` field accepts any JSON object.  If the key already exists,\nits value is replaced.",
        "operationId": "set_setting_api_v1_settings__key__put",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The created or updated setting",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AllowlistEntryCreate": {
        "properties": {
          "domain": {
            "type": "string",
            "minLength": 1,
            "title": "Domain",
            "description": "Domain name (e.g. 'go.dev', 'docs.python.org')",
            "examples": [
              "go.dev"
            ]
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "title": "Name",
            "description": "Human-friendly label for the domain",
            "examples": [
              "Go Official Blog"
            ]
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "Optional description of what this source provides"
          }
        },
        "type": "object",
        "required": [
          "domain",
          "name"
        ],
        "title": "AllowlistEntryCreate",
        "description": "Add a domain to the reading allowlist.\n\nOnly domains on the allowlist will be used for reading recommendations."
      },
      "AllowlistEntryResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "Entry ID"
          },
          "domain": {
            "type": "string",
            "title": "Domain",
            "description": "Allowlisted domain name"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Human-friendly label"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "Optional description"
          },
          "is_default": {
            "type": "boolean",
            "title": "Is Default",
            "description": "Whether this is a built-in default entry (cannot be deleted)"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "When the entry was added"
          }
        },
        "type": "object",
        "required": [
          "id",
          "domain",
          "name",
          "description",
          "is_default",
          "created_at"
        ],
        "title": "AllowlistEntryResponse",
        "description": "A domain on the reading recommendation allowlist."
      },
      "AllowlistEntryUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name",
            "description": "Updated label (null to keep current)"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "Updated description (null to keep current)"
          }
        },
        "type": "object",
        "title": "AllowlistEntryUpdate",
        "description": "Update an existing allowlist entry's name or description."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MessageResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable result message",
            "examples": [
              "Entry deleted"
            ]
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "MessageResponse",
        "description": "Generic message response for operations that return no entity."
      },
...
      "ReadingRecommendationResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "Recommendation ID"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "description": "Title of the recommended article or resource",
            "examples": [
              "Understanding Go's Concurrency Model"
            ]
          },
          "url": {
            "type": "string",
            "title": "Url",
            "description": "Full URL to the recommended resource",
            "examples": [
              "https://go.dev/blog/concurrency-is-not-parallelism"
            ]
          },
          "source_domain": {
            "type": "string",
            "title": "Source Domain",
            "description": "Domain the URL belongs to (must be on the allowlist)",
            "examples": [
              "go.dev"
            ]
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "Brief AI-generated summary of why this was recommended"
          },
          "topic_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic Id",
            "description": "Knowledge Profile topic this recommendation relates to"
          },
          "recommendation_type": {
            "$ref": "#/components/schemas/ReadingRecommendationType",
            "description": "Type of recommendation (deepening, broadening, refresher)"
          },
          "batch_date": {
            "type": "string",
            "format": "date",
            "title": "Batch Date",
            "description": "The weekly batch date this recommendation belongs to"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "When the recommendation was generated"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "Last modification timestamp"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "url",
          "source_domain",
          "description",
          "topic_id",
          "recommendation_type",
          "batch_date",
          "created_at",
          "updated_at"
        ],
        "title": "ReadingRecommendationResponse",
        "description": "A curated reading recommendation generated from the Knowledge Profile."
      },
      "ReadingRecommendationType": {
        "type": "string",
        "enum": [
          "next_frontier",
          "weak_spot",
          "deep_dive"
        ],
        "title": "ReadingRecommendationType"
      },
      "SettingResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "Setting ID"
          },
          "key": {
            "type": "string",
            "title": "Key",
            "description": "Unique setting key",
            "examples": [
              "quiz_question_count"
            ]
          },
          "value": {
            "additionalProperties": true,
            "type": "object",
            "title": "Value",
            "description": "Setting value as a JSON object"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "When the setting was last changed"
          }
        },
        "type": "object",
        "required": [
          "id",
          "key",
          "value",
          "updated_at"
        ],
        "title": "SettingResponse",
        "description": "A single application setting (key-value pair)."
      },
      "SettingUpdate": {
        "properties": {
          "value": {
            "additionalProperties": true,
            "type": "object",
            "title": "Value",
            "description": "Setting value as a JSON object",
            "examples": [
              {
                "model": "anthropic/claude-sonnet-4",
                "temperature": 0.7
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "value"
        ],
        "title": "SettingUpdate",
        "description": "Set a configuration value."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  },
...
}